+++ flake.nix
@@ -21,7 +21,7 @@
       {
         packages.default = pkgs.rustPlatform.buildRustPackage {
           pname = "notion-sync";
-          version = "0.2.1";
+          version = "0.3.0";
           src = ./.;
           cargoLock.lockFile = ./Cargo.lock;